pao:resources: move `getMachineConfigPoolByProfile` to `resources`
This is needed so we could use it outside the controller package in
future implementations.
