+++ pkg/performanceprofile/controller/performanceprofile/resources/resources.go
@@ -3,12 +3,14 @@ package resources
 import (
 	"context"
 	"encoding/json"
+	"fmt"
 	"reflect"
 
 	nodev1 "k8s.io/api/node/v1"
 	apiequality "k8s.io/apimachinery/pkg/api/equality"
 	"k8s.io/apimachinery/pkg/api/errors"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
+	"k8s.io/apimachinery/pkg/labels"
 	"k8s.io/apimachinery/pkg/types"
 	"k8s.io/klog"
 	kubeletconfigv1beta1 "k8s.io/kubelet/config/v1beta1"
@@ -16,6 +18,7 @@ import (
 
 	apiconfigv1 "github.com/openshift/api/config/v1"
 	mcov1 "github.com/openshift/api/machineconfiguration/v1"
+	performancev2 "github.com/openshift/cluster-node-tuning-operator/pkg/apis/performanceprofile/v2"
 	tunedv1 "github.com/openshift/cluster-node-tuning-operator/pkg/apis/tuned/v1"
 )
 
@@ -341,4 +344,56 @@ func DeleteRuntimeClass(ctx context.Context, cli client.Client, name string) err
 		return err
 	}
 	return cli.Delete(ctx, runtimeClass)
+}
+
+func GetMachineConfigPoolByProfile(ctx context.Context, client client.Client, profile *performancev2.PerformanceProfile) (*mcov1.MachineConfigPool, error) {
+	nodeSelector := labels.Set(profile.Spec.NodeSelector)
+
+	mcpList := &mcov1.MachineConfigPoolList{}
+	if err := client.List(ctx, mcpList); err != nil {
+		return nil, err
+	}
+
+	filteredMCPList := filterMCPDuplications(mcpList.Items)
+
+	var profileMCPs []*mcov1.MachineConfigPool
+	for i := range filteredMCPList {
+		mcp := &mcpList.Items[i]
+
+		if mcp.Spec.NodeSelector == nil {
+			continue
+		}
+
+		mcpNodeSelector, err := metav1.LabelSelectorAsSelector(mcp.Spec.NodeSelector)
+		if err != nil {
+			return nil, err
+		}
+
+		if mcpNodeSelector.Matches(nodeSelector) {
+			profileMCPs = append(profileMCPs, mcp)
+		}
+	}
+
+	if len(profileMCPs) == 0 {
+		return nil, fmt.Errorf("failed to find MCP with the node selector that matches labels %q", nodeSelector.String())
+	}
+
+	if len(profileMCPs) > 1 {
+		return nil, fmt.Errorf("more than one MCP found that matches performance profile node selector %q", nodeSelector.String())
+	}
+
+	return profileMCPs[0], nil
+}
+
+func filterMCPDuplications(mcps []mcov1.MachineConfigPool) []mcov1.MachineConfigPool {
+	var filtered []mcov1.MachineConfigPool
+	items := map[string]mcov1.MachineConfigPool{}
+	for _, mcp := range mcps {
+		if _, exists := items[mcp.Name]; !exists {
+			items[mcp.Name] = mcp
+			filtered = append(filtered, mcp)
+		}
+	}
+
+	return filtered
 }
